# tests/e2e_api/swagger.json
{
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "AuthUser": {
      "properties": {
        "id": {
          "example": "e2e-admin",
          "type": "string"
        },
        "role": {
          "example": "admin",
          "type": "string"
        },
        "username": {
          "example": "admin",
          "type": "string"
        }
      },
      "required": [
        "id",
        "username",
        "role"
      ],
      "type": "object"
    },
...
    "ErrorResponse": {
      "properties": {
        "error": {
          "example": "user not found",
          "type": "string"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
    "LoginRequest": {
      "properties": {
        "password": {
          "example": "admin123",
          "type": "string"
        },
        "username": {
          "example": "admin",
          "type": "string"
        }
      },
      "required": [
        "username",
        "password"
      ],
      "type": "object"
    },
    "LoginResponse": {
      "properties": {
        "token": {
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/AuthUser"
        }
      },
      "required": [
        "token",
        "user"
      ],
      "type": "object"
    },
...
    "User": {
      "properties": {
        "active": {
          "example": true,
          "type": "boolean"
        },
        "createdAt": {
          "example": "2026-01-01T00:00:00Z",
          "format": "date-time",
          "type": "string"
        },
        "email": {
          "example": "alice@example.test",
          "type": "string"
        },
        "id": {
          "example": "1",
          "type": "string"
        },
        "name": {
          "example": "Alice",
          "type": "string"
        },
        "role": {
          "example": "tester",
          "type": "string"
        },
        "updatedAt": {
          "example": "2026-01-01T00:00:00Z",
          "format": "date-time",
          "type": "string"
        }
      },
      "required": [
        "id",
        "name",
        "email",
        "role",
        "active",
        "createdAt",
        "updatedAt"
      ],
      "type": "object"
    }
  },
  "host": "localhost:18080",
  "info": {
    "description": "A local target API service for importing into the automation test platform.",
    "title": "Autotest E2E Demo API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [
          {
            "description": "Login credentials",
            "in": "body",
            "name": "credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "login response",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "400": {
            "description": "invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Login and get JWT",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "description": "Optional role filter",
            "in": "query",
            "name": "role",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "users",
            "schema": {
              "items": {
                "$ref": "#/definitions/User"
              },
              "type": "array"
            }
          },
          "401": {
            "description": "unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List users",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "AdminBearerAuth": {
      "description": "Use the value returned by /api/v1/admin/auth/login as: Bearer \u003ctoken\u003e",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    },
    "BearerAuth": {
      "description": "Use the value returned by /api/v1/auth/login as: Bearer \u003ctoken\u003e",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}